tests for enumerable collections and arrays pass
